[FIX]Libft/ft_demonize: Fix bad cast for sysconf return

+++ libft/srcs/deamonize/ft_demonize.c
@@ -6,8 +6,8 @@
 
 static inline void	close_all_fds()
 {
-	int	maxfd;
-	int	fd;
+	long	maxfd;
+	int		fd;
 
 	maxfd = sysconf(_SC_OPEN_MAX);
 	if (maxfd == -1)
